refactor(Exercise): move rating part from DetailsPanel to the exercise description
